requirements.txt
# Core ML and data processing dependencies
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0

# Testing framework
hypothesis>=6.82.0
pytest>=7.4.0

# Additional utilities
scikit-learn>=1.3.0
Pillow>=10.0.0
tqdm>=4.65.0
pyyaml>=6.0